Move earworm table to dynamic database
