+++ SKNIBot/Core/Database/DynamicDBContext.cs
@@ -14,6 +14,8 @@ namespace SKNIBot.Core.Database
         public virtual DbSet<AssignRole> AssignRoles { get; set; }
         public virtual DbSet<WelcomeMessage> WelcomeMessages { get; set; }
 
+        public virtual DbSet<SpotifyEarWorm> SpotifyEarWorms { get; set; }
+
         public DynamicDBContext() : base(GetOptions("Data Source=DynamicDatabase.sqlite"))
         {
