+++ SKNIBot/Core/Database/StaticDBContext.cs
@@ -18,8 +18,6 @@ namespace SKNIBot.Core.Database
         public virtual DbSet<HangmanCategory> HangmanCategories { get; set; }
         public virtual DbSet<HangmanWord> HangmanWords { get; set; }
 
-        public virtual DbSet<SpotifyEarWorm> SpotifyEarWorms { get; set; }
-
         public virtual DbSet<Role> Roles { get; set; }
 
         public StaticDBContext() : base(GetOptions("Data Source=StaticDatabase.sqlite"))
